+++ example/lib/widgets/configuration.dart
@@ -42,12 +42,8 @@ class _ConfigurationState extends State<Configuration> with ShowDialog {
   }
 
   final Map<String, Object> initialVisitorContext = {
-    "isVipClient": true,
-    "qa_getflag": true,
-    "bucketingKey": "condition1",
-    "QA": true,
-    "qaKeyString": "beta",
-    "qaKeyNumber": 2222
+    "customer": "QA",
+    "country": "AE"
   };
 
   Map<String, Object> visitorContext = {};
@@ -110,10 +106,25 @@ class _ConfigurationState extends State<Configuration> with ShowDialog {
           }
         })
         .withTimeout(int.tryParse(timeoutController.text) ?? defaultTimeout)
+        .withOnVisitorExposed((exposedUser, fromFlag) {
+          print("L'utilisateur ayant été exposé ");
+          print(fromFlag.toJson().toString());
+          print(exposedUser.toJson().toString());
+        })
         .build();
     Flagship.start(envIdController.text, apiKeyController.text, config: config);
   }
 
+  // .withUserExposureCallback((exposedUser, exposedFlag) {
+  //   ////
+  //   ///
+  //   ///
+
+  //   // -----------------------
+  //   print("Exposed flag value are : ${exposedFlag.toJson().toString()}");
+  //   print("Exposed flag value are : ${exposedUser.toJson().toString()}");
+  //   // -----------------------
+  // })
 // Change Mode
   _changeMode() {
     setState(() {