+++ lib/api/tracking_manager.dart
@@ -7,6 +7,7 @@ import 'package:flagship/hits/activate.dart';
 import 'package:flagship/flagship_version.dart';
 import 'package:flagship/utils/constants.dart';
 import 'package:flagship/utils/logger/log_manager.dart';
+import 'package:flutter/foundation.dart';
 import 'service.dart';
 import 'package:http/http.dart' as http;
 
@@ -37,7 +38,7 @@ class TrackingManager {
   }
 
   // later add code error in the future
-  Future<void> sendActivate(Activate activateHit) async {
+  Future<int> sendActivate(Activate activateHit) async {
     /// Create url
     String urlString = Endpoints.DECISION_API + Endpoints.ACTIVATION;
     var response = await _service.sendHttpRequest(
@@ -45,12 +46,14 @@ class TrackingManager {
         timeoutMs: TIMEOUT_REQUEST);
     switch (response.statusCode) {
       case 200:
+      case 201:
       case 204:
         Flagship.logger(Level.INFO, ACTIVATE_SUCCESS);
         break;
       default:
         Flagship.logger(Level.ERROR, ACTIVATE_FAILED);
     }
+    return response.statusCode;
   }
 
   /// Send Hit
@@ -58,7 +61,8 @@ class TrackingManager {
     /// Create url
     String urlString = Endpoints.ARIANE;
     try {
-      var response = await _service.sendHttpRequest(RequestType.Post, urlString, fsHeader, jsonEncode(pHit.bodyTrack),
+      var response = await _service.sendHttpRequest(
+          RequestType.Post, urlString, fsHeader, jsonEncode(pHit.bodyTrack),
           timeoutMs: TIMEOUT_REQUEST);
       switch (response.statusCode) {
         case 200:
@@ -70,8 +74,14 @@ class TrackingManager {
           Flagship.logger(Level.ERROR, HIT_FAILED);
       }
     } catch (error) {
-      Flagship.logger(Level.EXCEPTIONS, EXCEPTION.replaceFirst("%s", "$error") + urlString);
+      Flagship.logger(
+          Level.EXCEPTIONS, EXCEPTION.replaceFirst("%s", "$error") + urlString);
       Flagship.logger(Level.ERROR, HIT_FAILED);
     }
+  }
+
+  @visibleForTesting
+  void setService(Service pService) {
+    _service = pService;
   }
 }