+++ lib/visitor/strategy/panic_strategy.dart
@@ -1,5 +1,6 @@
 import 'package:flagship/flagship.dart';
 import 'package:flagship/hits/hit.dart';
+import 'package:flagship/model/modification.dart';
 import 'package:flagship/utils/constants.dart';
 import 'package:flagship/utils/logger/log_manager.dart';
 import 'package:flagship/visitor.dart';
@@ -11,6 +12,11 @@ class PanicStrategy extends DefaultStrategy {
 
   @override
   Future<void> activateModification(String key) async {
+    Flagship.logger(Level.INFO, PANIC_ACTIVATE);
+  }
+
+  @override
+  Future<void> activateFlag(Modification pFlag) async {
     Flagship.logger(Level.INFO, PANIC_ACTIVATE);
   }
 
@@ -27,7 +33,8 @@ class PanicStrategy extends DefaultStrategy {
 
   @override
   Map<String, Object>? getModificationInfo(String key) {
-    Flagship.logger(Level.ERROR, PANIC_MODIFICATION_INFO.replaceFirst("%s", key));
+    Flagship.logger(
+        Level.ERROR, PANIC_MODIFICATION_INFO.replaceFirst("%s", key));
     return null;
   }
 
@@ -49,5 +56,9 @@ class PanicStrategy extends DefaultStrategy {
   @override
   unAuthenticateVisitor() {
     Flagship.logger(Level.INFO, PANIC_UNAUTHENTICATE);
+  }
+
+  void onExposure(Modification pModification) {
+    Flagship.logger(Level.INFO, PANIC_ACTIVATE);
   }
 }