+++ test/flag_test.dart
@@ -1,11 +1,10 @@
 import 'dart:convert';
 
 import 'package:flagship/api/service.dart';
+import 'package:flagship/api/tracking_manager.dart';
 import 'package:flagship/decision/api_manager.dart';
 import 'package:flagship/flagship.dart';
-import 'package:flagship/flagshipContext/flagship_context_manager.dart';
 import 'package:flagship/flagship_config.dart';
-import 'package:flagship/flagship_version.dart';
 import 'package:flagship/model/flag.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart';
@@ -20,35 +19,45 @@ import 'package:http/http.dart' as http;
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
   SharedPreferences.setMockInitialValues({});
-  Map<String, String> fsHeaders = {
-    "x-api-key": "apiKey",
-    "x-sdk-client": "flutter",
-    "x-sdk-version": FlagshipVersion,
-    "Content-type": "application/json"
-  };
-
-  Map<String, dynamic> presetContext = FlagshipContextManager.getPresetContextForApp();
-  Map<String, dynamic> jsonData = {"visitorId": "flagVisitor", "context": presetContext, "trigger_hit": false};
-  Object data = json.encode(jsonData);
-  // Object data = json.encode({"visitorId": "flagVisitor", "context": {}, "trigger_hit": false});
   MockService fakeService = MockService();
   ApiManager fakeApi = ApiManager(fakeService);
 
-  String fakeResponse = await ToolsTest.readFile('test_resources/decisionApi.json') ?? "";
-  when(fakeService.sendHttpRequest(RequestType.Post,
-          'https://decision.flagship.io/v2/bkk9glocmjcg0vtmdlrr/campaigns/?exposeAllKeys=true', fsHeaders, data,
+  String fakeResponse =
+      await ToolsTest.readFile('test_resources/decisionApi.json') ?? "";
+  when(fakeService.sendHttpRequest(
+          RequestType.Post,
+          'https://decision.flagship.io/v2/bkk9glocmjcg0vtmdlrr/campaigns/?exposeAllKeys=true',
+          any,
+          any,
           timeoutMs: TIMEOUT))
       .thenAnswer((_) async {
     return http.Response(fakeResponse, 200);
   });
 
-  FlagshipConfig config = ConfigBuilder().withTimeout(TIMEOUT).build();
+  when(fakeService.sendHttpRequest(RequestType.Post,
+          'https://decision.flagship.io/v2/activate', any, any,
+          timeoutMs: 60000))
+      .thenAnswer((_) async {
+    return http.Response("fakeResponse", 200);
+  });
+
+  TrackingManager fakeTracking = TrackingManager();
+  fakeTracking.setService(fakeService);
+
+  FlagshipConfig config = ConfigBuilder()
+      .withTimeout(TIMEOUT)
+      .withOnVisitorExposed((exposedUser, exposedFlag) {
+    expect(exposedFlag.metadata().campaignId, "bsffhle242b2l3igq4dg");
+    expect(exposedFlag.metadata().variationGroupId, "bsffhle242b2l3igq4egaa");
+    expect(exposedFlag.metadata().variationId, "bsffhle242b2l3igq4f0");
+  }).build();
   config.decisionManager = fakeApi;
-  Flagship.start("bkk9glocmjcg0vtmdlrr", "apiKey", config: config);
-  var v1 = Flagship.newVisitor("flagVisitor").build();
+  await Flagship.start("bkk9glocmjcg0vtmdlrr", "apiKey", config: config);
 
   test("Test Flag class", (() async {
-    v1.fetchFlags().whenComplete(() {
+    var v1 = Flagship.newVisitor("flagVisitor").build();
+    v1.trackingManager = fakeTracking;
+    v1.fetchFlags().whenComplete(() async {
       // List of flags
       List<Map<String, dynamic>> listEntry = [
         // Correct Flag
@@ -59,9 +68,27 @@ void main() async {
           "existingFlag": true,
           "shouldHaveMetadata": true
         },
-        {"key": "key_B", "dfltValue": 2.14, "expectedValue": 3.14, "existingFlag": true, "shouldHaveMetadata": true},
-        {"key": "key_C", "dfltValue": 4, "expectedValue": 2, "existingFlag": true, "shouldHaveMetadata": true},
-        {"key": "key_D", "dfltValue": false, "expectedValue": true, "existingFlag": true, "shouldHaveMetadata": true},
+        {
+          "key": "key_B",
+          "dfltValue": 2.14,
+          "expectedValue": 3.14,
+          "existingFlag": true,
+          "shouldHaveMetadata": true
+        },
+        {
+          "key": "key_C",
+          "dfltValue": 4,
+          "expectedValue": 2,
+          "existingFlag": true,
+          "shouldHaveMetadata": true
+        },
+        {
+          "key": "key_D",
+          "dfltValue": false,
+          "expectedValue": true,
+          "existingFlag": true,
+          "shouldHaveMetadata": true
+        },
         // array
         {
           "key": "array",
@@ -108,20 +135,21 @@ void main() async {
           expect(metadata.variationGroupId, "");
           expect(metadata.variationId, "");
           expect(metadata.isReference, false);
-          expect(metadata.slug, "");
+          expect(metadata.slug, null);
           expect(metadata.campaignType, "");
         }
         // Check lentgh for metedata json
         expect(myFlag.metadata().toJson().keys.length, 6);
         // Expose
-        myFlag.userExposed();
+        await myFlag.userExposed();
       }
     });
   }));
 
   test("Flag with bad type", () {
-    v1.fetchFlags().whenComplete(() {
-      Flag myFlag = v1.getFlag("key_A", 3.14);
+    var v2 = Flagship.newVisitor("flagVisitor").build();
+    v2.fetchFlags().whenComplete(() {
+      Flag myFlag = v2.getFlag("key_A", 3.14);
       expect(myFlag.value(), 3.14);
       expect(myFlag.exists(), true);
       FlagMetadata metadata = myFlag.metadata();
@@ -129,14 +157,15 @@ void main() async {
       expect(metadata.variationGroupId, "");
       expect(metadata.variationId, "");
       expect(metadata.isReference, false);
-      expect(metadata.slug, "");
+      expect(metadata.slug, null);
       expect(metadata.campaignType, "");
     });
   });
 
   test("Flag with null as value", () {
-    v1.fetchFlags().whenComplete(() {
-      Flag myFlag = v1.getFlag("keyNull", "nullValue");
+    var v3 = Flagship.newVisitor("flagVisitor").build();
+    v3.fetchFlags().whenComplete(() {
+      Flag myFlag = v3.getFlag("keyNull", "nullValue");
       expect(myFlag.value(), "nullValue");
       expect(myFlag.exists(), true);
       FlagMetadata metadata = myFlag.metadata();
@@ -150,8 +179,9 @@ void main() async {
   });
 
   test("Flag value & default value = null", () {
-    v1.fetchFlags().whenComplete(() {
-      Flag myFlag = v1.getFlag("keyNull", null);
+    var v4 = Flagship.newVisitor("flagVisitor").build();
+    v4.fetchFlags().whenComplete(() {
+      Flag myFlag = v4.getFlag("keyNull", null);
       expect(myFlag.value(), null);
       expect(myFlag.exists(), true);
       FlagMetadata metadata = myFlag.metadata();
@@ -161,6 +191,20 @@ void main() async {
       expect(metadata.isReference, true);
       expect(metadata.slug, "flutter");
       expect(metadata.campaignType, "ab");
+    });
+  });
+
+  test("Failed Activate Flag", () {
+    when(fakeService.sendHttpRequest(RequestType.Post,
+            'https://decision.flagship.io/v2/activate', any, any,
+            timeoutMs: 60000))
+        .thenAnswer((_) async {
+      return http.Response("fakeResponse", 400);
+    });
+    var v5 = Flagship.newVisitor("flagVisitor").build();
+    v5.fetchFlags().whenComplete(() async {
+      Flag myFlag = v5.getFlag("key_A", "12");
+      myFlag.userExposed();
     });
   });
 }